requirements.txt
pandas==2.2.2
sqlalchemy==2.0.30
oracledb==2.3.0
pyyaml==6.0.1
